+++ CHANGELOG.md
@@ -53,8 +53,8 @@
 - Memory efficiency improvements - split chapter content is only loaded when accessed
 
 # 5.1.0
-- Added EpubReader.readBookWithSplitChapters to split long chapters (>5000 words) into smaller parts
-- Added getChaptersWithSplitting to get a list of chapters while taking into account splitting longer chapters (>5000 words)
+- Added EpubReader.readBookWithSplitChapters to split long chapters (>3000 words) into smaller parts
+- Added getChaptersWithSplitting to get a list of chapters while taking into account splitting longer chapters (>3000 words)
 
 # 5.0.5
 - Ran dart formatter to improve package score